Add files via upload
This text file provides a record of the user's responses to prompts for user input made by the BIGS903 software application. The record shows the following. The user specified a subset of game rules. The subset of game rules specified the following. The dealer deals from a six deck shoe. The composition of each deck is the standard composition. The standard composition of is thirteen ranks of each of four suits plus two jokers.  Any hand that includes a joker has a numerical value of twenty-seven.